requirements.txt
playwright==1.44.0
flask>=3.0
openai==1.30.0
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
pytest==8.2.0
pytest-asyncio==0.23.7
